requirements.txt
fastapi~=0.104.1
uvicorn[standard]~=0.23.2

pypdf~=3.17.0
python-docx~=1.0.1
python-pptx~=0.6.22
openpyxl~=3.1.2
azure-cognitiveservices-speech~=1.32.1
